+++ fastapi_server/app/config.py
@@ -105,6 +105,8 @@ class Config(DataRobotAppFrameworkBaseSettings):
     # The number of characters to stream before persisting
     minimal_chunks_to_persist: int = 5000
 
+    profiling_enabled: bool = False
+
     application_id: str | None = None
 
     otel_entity_id: str = ""